used incorrect datetime parsing string (12h vs 24h)
Accidentally used the wrong format string for date parsing.
the string I used was for a 12 hour time, whereas the
input is in 24 hour time.  This was not discovered during
testing because the test data I was using was timestamped
at 01:something:something, i.e. it wouldn't trigger the
bug; it wasn't until actual demo time, when UTZc time was
>12 that the bug showed up.